# Vectorise string.Replace(char, char)