* Replace bits of existing unsafe code to IndexOf, which is already vectorised
* Vectorise the replacement bit

I will PR more tests/benchmarks that test both the non-vectorised and vectorised paths in corefx/performance repo. (In the meantime, I have verified that the tests pass locally.)

Also, I am aware that the closure of the repo for consolidation is due; I am happy to open the PR again once the repo is open, if this cannot be merged before the closure.

Tested with four cases (no characters replaced / some characters replaced, with two different strings), and the results seem to suggest slight slowdown for strings that are too small to be vectorised & has something to be replaced (here it appears to be around 20% slower), but for longer strings it seems to be better:
```
summary:
better: 2, geomean: 1.293
worse: 1, geomean: 1.193
total diff: 3
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Tests.Perf_String.Replace_Char(text: "Hello", oldChar: 'l', newChar: '!') |      1.19 |             9.44 |            11.27 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_String.Replace_Char(text: "This is a very nice sentence", oldChar: 'i', newChar: 'I') |      1.33 |             9.36 |             7.05 |         |
| System.Tests.Perf_String.Replace_Char(text: "This is a very nice sentence", oldChar: 'z', newChar: 'y') |      1.26 |            22.62 |            17.95 | several?|