@GrabYourPitchforks 2019-11-11
Why not use the normal `Vector<T>` APIs for this? As far as I can tell we're not relying on specific x86 intrinsic instructions. Using `Vector<T>` also allows this code to "just work" on ARM platforms.