@Gnbrkm41 2019-11-12
New perf numbers after running it on an AVX2 supported machine:
```
C:\Users\gotos\source\repos\performance\src\tools\ResultsComparer>dotnet run --base C:\Users\gotos\desktop\replace-before --diff C:\users\gotos\desktop\replace-after --threshold 5%
summary:
better: 4, geomean: 2.626
worse: 1, geomean: 1.204
total diff: 5
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| `"Hello".Replace('l', '!')` |      1.20 |            10.49 |            12.63 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| `"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut id tempor metus.".Replace('z', 'y')` |      5.19 |            46.50 |             8.97 |         |
| `"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut id tempor metus.".Replace('e', 'E')` |      2.79 |            66.68 |            23.93 |         |
| `"This is a very nice sentence"`.Replace('z', 'y')` |      2.21 |            14.65 |             6.62 |         |
|  `"This is a very nice sentence"`.Replace('i', 'I')` |      1.49 |            28.97 |            19.47 | several?|